Add PostgreSQL database configuration in Django settings
The "dummy" database configuration in Django settings has been replaced with a PostgreSQL database configuration using the dj-database-url package. This change is necessary to ensure this Django app can operate over HTTP, which is required in serverless environments like Vercel. The related package has been added to the project's requirements.